requirements-dev.txt
# Core dependencies
protobuf>=3.20.0,<4.0.0
streamlit==1.27.0
pandas>=2.0.0,<3.0.0
pandas-ta==0.3.14b0
numpy>=1.21.0,<1.24.0
requests==2.31.0
requests-oauthlib==1.3.1
plotly>=4.12.0,<6.0.0
torch==2.0.1
scikit-learn==1.3.2
python-dateutil>=2.8.2,<3.0.0

# Trading & Backtesting
backtrader>=1.9.78,<1.10.0
yfinance==0.2.30

# Alternative trading libraries
backtesting==0.3.3
vectorbt==0.27.2

# Data Visualization
matplotlib>=3.7.0,<4.0.0

# Development & Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.1.0; python_version >= "3.7"
flake8==6.0.0; python_version >= "3.7"
isort>=5.0.0,<6.0.0

# Type Checking
mypy==1.0.1
types-requests==2.31.0

# Validation
pydantic-settings>=2.0.0

python-dotenv